feat: Add loading spinner to activity asset in popup window
Added css based loading spinner to show user if cover image is loading. Currently only added and not implemented, meaning it is always visible and does not show any status right now.

fix: Correct argument passing for event based function calling

Added as first parameter to toggle_host_selection an event parameter, that is automatically passed in when setting it as event callback, what lead to unexpected behaviour before, because it was passed to force_close parameter. Set _event to null if you call it directly.